# frontend/types/openapi/oas30.yaml
openapi: "3.0.1"
info:
  title: "pseudopoll-rest-api"
  version: "2024-01-26T19:26:38Z"
servers:
- url: "https://api.pseudopoll.com"
paths:
  /polls/{pollId}/duration:
    patch:
      parameters:
      - name: "pollId"
        in: "path"
        required: true
        schema:
          type: "string"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePollDuration"
        required: true
      responses:
        "404":
          description: "404 response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "200":
          description: "200 response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdatePollDuration"
        "400":
          description: "400 response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "500 response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - pseudopoll-api-authorizer: []
      x-amazon-apigateway-request-validator: "update-poll-duration-validator"
      x-amazon-apigateway-integration:
        type: "aws_proxy"
        httpMethod: "POST"
        uri: "arn:aws:apigateway:us-east-2:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-2:981543223421:function:pseudopoll-update-poll-duration/invocations"
        passthroughBehavior: "when_no_match"
        timeoutInMillis: 29000
  /polls/{pollId}/archive:
    patch:
      parameters:
      - name: "pollId"
        in: "path"
        required: true
        schema:
          type: "string"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArchivePoll"
        required: true
      responses:
        "400":
          description: "400 response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "500 response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "204":
          description: "204 response"
          content: {}
      security:
      - pseudopoll-api-authorizer: []
      x-amazon-apigateway-request-validator: "archive-poll-validator"
      x-amazon-apigateway-integration:
        type: "aws_proxy"
        httpMethod: "POST"
        uri: "arn:aws:apigateway:us-east-2:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-2:981543223421:function:pseudopoll-archive-poll/invocations"
        passthroughBehavior: "when_no_match"
        timeoutInMillis: 29000
  /polls/{pollId}/{optionId}:
    post:
      parameters:
      - name: "pollId"
        in: "path"
        required: true
        schema:
          type: "string"
      - name: "optionId"
        in: "path"
        required: true
        schema:
          type: "string"
      responses:
        "202":
          description: "202 response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteAccepted"
      security:
      - pseudopoll-api-authorizer: []
      x-amazon-apigateway-integration:
        type: "aws"
        credentials: "arn:aws:iam::981543223421:role/pseudopoll-api-gateway-role"
        httpMethod: "POST"
        uri: "arn:aws:apigateway:us-east-2:sqs:path/pseudopoll-vote-queue"
        responses:
          ^2[0-9][0-9]:
            statusCode: "202"
            responseTemplates:
              application/json: "{\n  \"message\": \"Vote queued.\",\n  \"requestId\"\
                : \"$context.extendedRequestId\"\n}\n"
        requestParameters:
          integration.request.header.Content-Type: "'application/x-www-form-urlencoded'"
        requestTemplates:
          application/json: "#set($parameters = $input.path('$'))\n#set($parameters.pollId\
            \ = $input.params('pollId'))\n#set($parameters.optionId = $input.params('optionId'))\n\
            #set($parameters.userId = $context.authorizer.principalId)\n#set($parameters.userIp\
            \ = $input.params('x-user-ip'))\n#set($parameters.requestTimeEpoch = $context.requestTimeEpoch)\n\
            #set($parameters.requestId = $context.extendedRequestId)\n#set($body =\
            \ \"{\")\n#foreach($k in $parameters.keySet())\n#set($body = $body + \"\
            \"\"$k\"\": \"\"\" + $util.escapeJavaScript($parameters.get($k)) + \"\"\
            \", \")\n#end\n#set($body = $body + \"}\")\nAction=SendMessage&MessageBody=$body.replaceAll(\"\
            , }\", \"}\")\n"
        passthroughBehavior: "never"
        timeoutInMillis: 29000
  /polls/{pollId}:
    get:
      parameters:
      - name: "pollId"
        in: "path"
        required: true
        schema:
          type: "string"
      responses:
        "500":
          description: "500 response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "200":
          description: "200 response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poll"
        "403":
          description: "403 response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - pseudopoll-api-authorizer: []
      x-amazon-apigateway-integration:
        type: "aws_proxy"
        httpMethod: "POST"
        uri: "arn:aws:apigateway:us-east-2:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-2:981543223421:function:pseudopoll-get-poll/invocations"
        passthroughBehavior: "when_no_match"
        timeoutInMillis: 29000
  /public/polls/{pollId}:
    get:
      parameters:
      - name: "pollId"
        in: "path"
        required: true
        schema:
          type: "string"
      responses:
        "200":
          description: "200 response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poll"
        "401":
          description: "401 response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "500 response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: "403 response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-amazon-apigateway-integration:
        type: "aws_proxy"
        httpMethod: "POST"
        uri: "arn:aws:apigateway:us-east-2:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-2:981543223421:function:pseudopoll-get-poll/invocations"
        passthroughBehavior: "when_no_match"
        timeoutInMillis: 29000
  /polls:
    get:
      responses:
        "500":
          description: "500 response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "200":
          description: "200 response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyPolls"
        "403":
          description: "403 response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - pseudopoll-api-authorizer: []
      x-amazon-apigateway-integration:
        type: "aws_proxy"
        httpMethod: "POST"
        uri: "arn:aws:apigateway:us-east-2:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-2:981543223421:function:pseudopoll-my-polls/invocations"
        passthroughBehavior: "when_no_match"
        timeoutInMillis: 29000
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePoll"
        required: true
      responses:
        "400":
          description: "400 response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "500 response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "201":
          description: "201 response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poll"
      security:
      - pseudopoll-api-authorizer: []
      x-amazon-apigateway-request-validator: "create-poll-validator"
      x-amazon-apigateway-integration:
        type: "aws_proxy"
        httpMethod: "POST"
        uri: "arn:aws:apigateway:us-east-2:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-2:981543223421:function:pseudopoll-create-poll/invocations"
        passthroughBehavior: "when_no_match"
        timeoutInMillis: 29000
  /public/polls/{pollId}/{optionId}:
    post:
      parameters:
      - name: "pollId"
        in: "path"
        required: true
        schema:
          type: "string"
      - name: "optionId"
        in: "path"
        required: true
        schema:
          type: "string"
      responses:
        "202":
          description: "202 response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteAccepted"
      x-amazon-apigateway-integration:
        type: "aws"
        credentials: "arn:aws:iam::981543223421:role/pseudopoll-api-gateway-role"
        httpMethod: "POST"
        uri: "arn:aws:apigateway:us-east-2:sqs:path/pseudopoll-vote-queue"
        responses:
          ^2[0-9][0-9]:
            statusCode: "202"
            responseTemplates:
              application/json: "{\n  \"message\": \"Vote queued.\",\n  \"requestId\"\
                : \"$context.extendedRequestId\"\n}\n"
        requestParameters:
          integration.request.header.Content-Type: "'application/x-www-form-urlencoded'"
        requestTemplates:
          application/json: "#set($parameters = $input.path('$'))\n#set($parameters.pollId\
            \ = $input.params('pollId'))\n#set($parameters.optionId = $input.params('optionId'))\n\
            #set($parameters.userId = $context.authorizer.principalId)\n#set($parameters.userIp\
            \ = $input.params('x-user-ip'))\n#set($parameters.requestTimeEpoch = $context.requestTimeEpoch)\n\
            #set($parameters.requestId = $context.extendedRequestId)\n#set($body =\
            \ \"{\")\n#foreach($k in $parameters.keySet())\n#set($body = $body + \"\
            \"\"$k\"\": \"\"\" + $util.escapeJavaScript($parameters.get($k)) + \"\"\
            \", \")\n#end\n#set($body = $body + \"}\")\nAction=SendMessage&MessageBody=$body.replaceAll(\"\
            , }\", \"}\")\n"
        passthroughBehavior: "when_no_templates"
        timeoutInMillis: 29000
components:
  schemas:
    VoteAccepted:
      title: "Vote Accepted Schema"
      required:
      - "requestId"
      type: "object"
      properties:
        message:
          type: "string"
        requestId:
          type: "string"
    ArchivePoll:
      title: "Archive Poll Schema"
      required:
      - "value"
      type: "object"
      properties:
        value:
          type: "boolean"
    CreatePoll:
      title: "Create Poll Schema"
      required:
      - "duration"
      - "options"
      - "prompt"
      type: "object"
      properties:
        prompt:
          maxLength: 280
          minLength: 1
          type: "string"
          description: "The poll prompt"
        options:
          maxItems: 10
          minItems: 2
          uniqueItems: true
          type: "array"
          description: "The options to vote on"
          items:
            maxLength: 35
            minLength: 1
            type: "string"
        duration:
          maximum: 604800
          minimum: 60
          type: "integer"
          description: "The duration of the poll in seconds"
    Error:
      title: "Error schema"
      required:
      - "message"
      type: "object"
      properties:
        message:
          type: "string"
          description: "The error message"
        cause:
          type: "string"
          description: "The cause of the error"
    Poll:
      title: "Poll Schema"
      required:
      - "createdAt"
      - "duration"
      - "isArchived"
      - "options"
      - "pollId"
      - "prompt"
      - "userId"
      type: "object"
      properties:
        pollId:
          maxLength: 12
          minLength: 12
          type: "string"
        userId:
          type: "string"
        prompt:
          maxLength: 280
          minLength: 1
          type: "string"
          description: "The poll prompt text"
        options:
          maxItems: 10
          minItems: 2
          uniqueItems: true
          type: "array"
          description: "The options to vote on"
          items:
            required:
            - "isMyVote"
            - "optionId"
            - "pollId"
            - "text"
            - "updatedAt"
            - "votes"
            type: "object"
            properties:
              optionId:
                maxLength: 12
                minLength: 12
                type: "string"
              pollId:
                maxLength: 12
                minLength: 12
                type: "string"
              text:
                maxLength: 140
                minLength: 1
                type: "string"
                description: "The option text"
              updatedAt:
                type: "string"
                description: "The time of the last vote on this option"
              votes:
                minimum: 0
                type: "integer"
                description: "The number of votes for this option"
              isMyVote:
                type: "boolean"
                description: "Whether the current user has voted for this option"
        createdAt:
          type: "string"
          description: "The time the poll was created"
        duration:
          maximum: 604800
          minimum: 60
          type: "integer"
          description: "The duration of the poll in seconds"
        isArchived:
          type: "boolean"
          description: "Whether the poll is archived"
    UpdatePollDuration:
      title: "Update Poll Duration Schema"
      required:
      - "value"
      type: "object"
      properties:
        value:
          minimum: -1
          type: "number"
    MyPolls:
      title: "My Poll Schema"
      type: "array"
      items:
        required:
        - "createdAt"
        - "duration"
        - "isArchived"
        - "pollId"
        - "prompt"
        - "userId"
        type: "object"
        properties:
          pollId:
            maxLength: 12
            minLength: 12
            type: "string"
          userId:
            type: "string"
          prompt:
            maxLength: 280
            minLength: 1
            type: "string"
            description: "The poll prompt text"
          createdAt:
            type: "string"
            description: "The time the poll was created"
          duration:
            maximum: 604800
            minimum: 60
            type: "integer"
            description: "The duration of the poll in seconds"
          isArchived:
            type: "boolean"
            description: "Whether the poll is archived"
  securitySchemes:
    pseudopoll-api-authorizer:
      type: "apiKey"
      name: "Authorization"
      in: "header"
      x-amazon-apigateway-authtype: "custom"
      x-amazon-apigateway-authorizer:
        type: "token"
        authorizerUri: "arn:aws:apigateway:us-east-2:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-2:981543223421:function:pseudopoll-api-authorizer/invocations"
        authorizerCredentials: "arn:aws:iam::981543223421:role/api_gateway_authorizer_invocation_role"
        authorizerResultTtlInSeconds: 0
x-amazon-apigateway-gateway-responses:
  DEFAULT_4XX:
    responseParameters:
      gatewayresponse.header.Access-Control-Allow-Methods: "'OPTIONS'"
      gatewayresponse.header.Access-Control-Allow-Credentials: "'true'"
      gatewayresponse.header.Access-Control-Allow-Origin: "'*'"
      gatewayresponse.header.Access-Control-Allow-Headers: "'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token'"
  DEFAULT_5XX:
    responseParameters:
      gatewayresponse.header.Access-Control-Allow-Methods: "'OPTIONS'"
      gatewayresponse.header.Access-Control-Allow-Credentials: "'true'"
      gatewayresponse.header.Access-Control-Allow-Origin: "'*'"
      gatewayresponse.header.Access-Control-Allow-Headers: "'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token'"
    responseTemplates:
      application/json: "{\"message\":$context.error.messageString}"
x-amazon-apigateway-request-validators:
  archive-poll-validator:
    validateRequestParameters: false
    validateRequestBody: true
  create-poll-validator:
    validateRequestParameters: false
    validateRequestBody: true
  update-poll-duration-validator:
    validateRequestParameters: false
    validateRequestBody: true